Allow "null" for maps, beans, list
Until now, passing in "null" into the parser when it was expecting to
convert into a map, list or bean would result in a token error because
the first token would be be '{' or ']'.

While it makes sense in some cases that we don't want null values for
parsing results, it doesn't seem something that svenson should prevent,
as there might be other cases where we want "null" to mean absence
of something. Especially since we can represent the JSON "null" directly
as Java null. This allows clients not to special case null values that
might occur in their JSON communication.

This changes and tests the behavior to return null in this case.