Fix overloading issue
Fix issue 23 which turned out not to be PropertyDescriptor related,
mainly because we don't use PropertyDescriptor ;)

Instead the mistake was that when both the correct and its super method
would show up in object analysis, it could happen that the super method
overwrote the discovered overloaded method.

Now overwriting is only allowed when the overwriting method is the
overloaded method of the method to be overwritten and not vice versa.

Added JavaObjectSupport.isOveriding(Class,Class) and added a test
for it.