Only allow adders without settter
The fix in 1.4.2 fixing some adder cases could result in the adders
being a little too eager and "grab the control flow" even in some
cases where they can't handle the data but a setter could.

Still wanting to support the old adder feature, I tuned it down
so that an adder will only be found when no setter exists for
the same property. The reasoning being that an setter that is
in fact of a matching collection type can just as well be
used to parse data and a setter with mismatched type is not
a scenario we want to support.