Implement property priority.
Add a priority attribute to the @JSONProperty annotation to control order of JSON
properties on output.

While it makes no difference semantically, there are situations where it can
be either aesthetically pleasing or faster to have a certain property first.

The latter is the case when doing a JSON lookahead.

The priority values control the sort order of JSON properties. Higher values
come first, default is 0. Negative values are allowed, too.