Change escaping behavior for slashes
We always escaped slashes which apparently can cause problems with badly
written software elsewhere.

Now the main purpose of escaping slashes is to allow embedding of JSON
in Javascript / JSONP. The RFC 4627 does not require us to escape slashes.

Is there any purpose for escaping slashes other than that? If so it
elludes me.