Remove type converters from type analyzer caching
It was a bit idea to cache such a value with high variance among use-cases
together with class structure information that rarely changes, if ever.

We remove the evaluation of typeconverters at property info creation time
and move it back to runtime resolution. Thankfully, the API was still
kept the way we need it for that so there are no client code changes
necessary.

This fixes issue #27.