Support type-mapped object / literal variance
Parsing JSON structures with varying structures into Java object is a
special challenge given the nature of Java's type system and Java beans.

You end up with a "lowest common denominator" type in the properties you
need to have varied values, which usually means having Object properties.

There seems to be one way to support such setups in a meaningful viable
way.

We can allow an type-upgrading style of type matching.

If you take a look at org.svenson.ComplexLookaheadTestCase you see two
kinds of way of doing that. One is to use a simple Object property but
annotate it with @JSONTypeHint() with a specific base type we use in a
ClassNameBasedTypeMapper setup.

public class PropertyVarianceContainer
{
    private Object foo;

    public Object getFoo()
    {
        return foo;
    }

    @JSONTypeHint(Base.class)
    public void setFoo(Object foo)
    {
        this.foo = foo;
    }
}

Here we have a parameter formally declared as Object but with a typehint.
This now leads of course to any kind of markup being accepted for this
property. So any primitive and map and list will just be set here. In
addition to that this commit ensures that the type mapper correctly
produces typed subclasses of our base here.

The same works for Maps and Lists:

public class MapVarianceContainer
{
    private Map<String,Object> props;

    public Map<String, Object> getProps()
    {
        return props;
    }

    @JSONTypeHint(Foo.class)
    public void setProps(Map<String, Object> props)
    {
        this.props = props;
    }
}

public class ListVarianceContainer
{
    private List<Object> props;

    public List<Object> getProps()
    {
        return props;
    }

    @JSONTypeHint(Foo.class)
    public void setProps(List<Object> props)
    {
        this.props = props;
    }

again we declare the Map<String,Object>/List<Ojbject> and add the TypeHint to
get the type conversion we want.

Due to the accept-all nature of Object the values set can of course be anything. You
have to make sure to manually validate the values and accept only those that
make sense in your domain.

It also makes a lot of sense to provide better typed accessors for the variants you
want to accept.