Implement @JSONParameter
Add an @JSONParameter annotation to be able to parse into immutable java
classes via an annotated constructor

Exactly one constructor of a class that has no setter or adder can be
annotated with @JSONParameter annotations.

The parser will use a new class DelayedConstructor which implements
DynamicProperies as temporary storage for the JSON properties.

The placeholder uses the object array in the end used as constructor
parameters as storage for the JSON property values.

It will error if a property is set for which there is no @JSONParameter
annotation.

Parameters that are not provided are substituted with null which will cause
an IllegalArgumentException on natively typed constructor parameters.

Test several scenarios of creating immutables, injecting immutables into
immutables and adding immutables.

The latter has also uncovered an old bug where the paramter type of an
adder method was not used as typehint.