Prevent AbstractTypeMapper NPE
Prevent an NPE caused by AbstractTypeMapper returning "null", that is
abstaining from the type decision.

This is useful e.g. to have an optional type discriminator present which
is generated by the system but not needed on reparsing if parse is
called with the correct type known by other means.

Add test.