Improve JSONifying behavior of cyclic structures.
One of the things svenson does not do for you is to take care of cyclic
structure. Instead it expects you to make sure that you
@JSONPropery(ignore=true) the right back references to solve the issue.

It would just fail with a StackoverflowError when you fail to do so or
when you use types that are cyclic by nature.

See also https://github.com/fforw/svenson/issues/1 (not actually the
first issue, just the first one since we moved to github)

One reason why the StackOverflowError is so annoying lies in the its
very nature of being a stack overflow and its stack trace being a
loooong chain of repetitive method invocations.

Thankfully there is one easy fix we can do right now without deciding on
the more complicated issues outline in the bug.

We already do the JSONifying in a private recursive method which is
called from public non-recursive methods. Which means that if we catch
the stack overflow error there, we can wrap it in another exception
whose stack trace then shows us the entry point into the JSONification
process, greatly aiding debugging.