JSONBUilder: Allow using outside sink
Make it possible to create a JSON builder based on an existing sink.
This can be highly useful if you use the JSONBuilder with a
SinkAwareJSONifier.

This introduces a new error condition where we can give e.g. a
WriterSink from which we cannot get the output back. In this case
.output() will throw an Exception.

Using alternative Sinks however means just that you shouldn't be
interested in the builder output but want to embed it into the
sink of a larger JSON generation. So this error case shouldn't cause too
much trouble.

We add a .closeAll() method to easily close all remaining open levels
as a helper for this case.

We add some very simple tests to test all of the now six JSONBuilder entry
points