JSONBuilder: Fix borked .include
He who writes the wrong code also writes non-sensical tests to prove
them right.

The include feature of course doesn't make sense as it was as it only
produced valid JSON in an element context :\

The structural anomaly to the other methods should have tipped me off.

Well.. now it works right with two Different methods on the builder

 JSONBuilder.includeProperty(String name, String json)

and

 JSONBuilder.includeElement(String json)

both validate their contexts and .includeProperty outputs the JSON
attribute key with the given name.