@JSONParameter: Allow ignored setters
When we introduced @JSONParameter(s) constructor parsing, we introduced
a rule in the JavaObjectSupport which disallowed any setters to be used
if the class uses constructor injection. Main motivation being both the
use-case of immutable classes as well as not wanting to deal with
conflicts in naming etc.

Now a system might have classes that are read from JSON sources without
modification on the Java side so they want to use @JSONParameter.

But the system tracks additional information about the JSON blob in
@JSONProperty(ignore=true) marked setters.

The class is no longer immutable, but we're not the immutability police
either. So we allow setters if they don't exist on the JSON level.

Adds two more test cases.