TypeAnalyzer: Make JSONPropertyInfo impls public
Make the two JSONPropertyInfo implementations in the project public so
it is either to extend on svenson JSON semantics with your own
extensions.

There is solid use cases in using the type knowledge gained by the
TypeAnalyzer, reusing the exact properties and property names and
settings it finds and extend that in your own project with e.g. extra
annotations. Custom automatic documentation generation, JSON-Schema
generation etc etc.