Modernize release plugins, suppress slf4j CVE
We take over some newer maven config from another project of main,
hopefully doing automated Sonatype OSS deploys from the get go.

The CVE on slf4j seems to have a) no reasonable solution yet and b)
refers to an EventData class that is not among our transitive
dependencies, so we suppress it.