Fix type checking in RecastUtil
Again, the most insidiuous errors prove to be the ones that
superficially work and therefore make you stop investigating.

All the type checks in RecastUtil where the wrong way around which works
fine as long as we're comparing collection types. But when the prop type
is Object and therefore assignable from everything it all breaks down.

Adds an additional test case with a bean with Object-typed property.