Reduce module footprint
As it was, we introduced a module dependency for "java.desktop" just for
the Introspector.decapitalize method. The whole current stack seems to
use no other class or method from it. This seems like a bad situation,
especially given how trivial that method is.

We create our own method for that purpose.

Also removes an unused exception wrapper method that referenced
java.beans.IntrospectionException.